requirements.txt
telethon>=1.36.0
discord.py>=2.3.0
requests>=2.31.0
rich>=13.0.0
python-dotenv>=1.0.0
pycryptodome>=3.19.0
pysocks>=1.7.0
fpdf2>=2.7.0
supabase>=2.0.0
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
aiohttp>=3.9.0
